delete %ENV entries after localizing them to bypass taint
Doing local $ENV{$var} localizes the environment variable, but also sets
it to undef, which for an env var means setting it to an empty string.
Some environment variables are problematic when set to an empty string.
For example, an empty PATH is equivalent to '.', the current directory,
which makes it vulnerable to anything in the current directory. Future
versions of perl will complain about a PATH set to an empty string or
undef when run under taint mode.

Deleting PATH after localizing it will avoid this problem. For
consistency, just delete all of the env vars.